Add spend-to-P2TR
Adds code to validate bech32m encoded taproot addresses
and scriptPubKeys. The underlying functions are in
python-bitcointx upcoming version 1.1.3.
This commit also updates the commit and compilation of
libsecp256k1 underlying library that is necessary for the
above and for spending and construction of taproot outputs,
as we may need in future.